refactor internal file structure
