+++ internal/base.hpp
@@ -1,0 +1,15 @@
+// Source Code: https://github.com/hit9/quadtree-pathfinding
+// License: BSD. Chao Wang, hit9[At]icloud.com.
+
+#ifndef QDPF_INTERNAL_BASE_HPP
+#define QDPF_INTERNAL_BASE_HPP
+
+namespace qdpf {
+namespace internal {
+
+const int inf = 0x3f3f3f3f;
+
+}  // namespace internal
+}  // namespace qdpf
+
+#endif
